requirements.txt
Flask
pymongo
python-dotenv
yt_dlp
werkzeug
gunicorn